requirements.txt
anthropic==0.49.0
chromadb==0.5.0
Flask==3.1.3
python-dotenv==1.2.2
pypdf==4.3.1
gunicorn==21.2.0
numpy==1.26.4
sentence-transformers==2.2.2
